examples/subscribe/requirements.txt
PyJWT>=1.7.1
sseclient>=0.0.26
